Update app layout to simplify interface
