+++ Views/Components/EntityViews/EntityGridView.swift
@@ -7,10 +7,9 @@ struct EntityGridView<T: Entity>: View {
 
     @State private var hoveredEntityID: UUID?
     @State private var isScrolling = false
-    @State private var scrollWorkItem: DispatchWorkItem?
-
+    
     private let columns = [
-        GridItem(.adaptive(minimum: ViewDefaults.gridArtworkSize, maximum: 200), spacing: 16)
+        GridItem(.adaptive(minimum: ViewDefaults.gridArtworkSize, maximum: ViewDefaults.gridArtworkSize + 40), spacing: 16)
     ]
 
     var body: some View {
@@ -40,32 +39,9 @@ struct EntityGridView<T: Entity>: View {
             }
             .padding(.horizontal, 14)
             .padding(.vertical, 6)
-            .background(
-                GeometryReader { geometry in
-                    Color.clear.preference(
-                        key: ScrollOffsetPreferenceKey.self,
-                        value: geometry.frame(in: .named("scroll")).origin.y
-                    )
-                }
-            )
         }
         .coordinateSpace(name: "scroll")
-        .onPreferenceChange(ScrollOffsetPreferenceKey.self) { _ in
-            handleScrollChange()
-        }
-    }
-    
-    private func handleScrollChange() {
-        isScrolling = true
-        hoveredEntityID = nil
-        
-        scrollWorkItem?.cancel()
-        
-        scrollWorkItem = DispatchWorkItem {
-            isScrolling = false
-        }
-        
-        DispatchQueue.main.asyncAfter(deadline: .now() + 0.2, execute: scrollWorkItem!)
+        .modifier(ScrollDetectionModifier(isScrolling: $isScrolling, hoveredEntityID: $hoveredEntityID))
     }
 
     @ViewBuilder
@@ -87,7 +63,128 @@ struct EntityGridView<T: Entity>: View {
     }
 }
 
-// MARK: - Entity Grid Item
+// MARK: - Cross-OS Scroll Detection
+
+private struct ScrollDetectionModifier: ViewModifier {
+    @Binding var isScrolling: Bool
+    @Binding var hoveredEntityID: UUID?
+    
+    func body(content: Content) -> some View {
+        if #available(macOS 15.0, *) {
+            content
+                .onScrollPhaseChange { _, newPhase in
+                    withAnimation(.none) {
+                        let wasScrolling = isScrolling
+                        isScrolling = newPhase == .interacting || newPhase == .decelerating
+                        
+                        if isScrolling && !wasScrolling {
+                            hoveredEntityID = nil
+                        }
+                    }
+                }
+        } else {
+            content
+                .background(
+                    ScrollDetectionView { isDetectedScrolling in
+                        if isDetectedScrolling != isScrolling {
+                            withAnimation(.none) {
+                                isScrolling = isDetectedScrolling
+                                if isScrolling {
+                                    hoveredEntityID = nil
+                                }
+                            }
+                        }
+                    }
+                )
+        }
+    }
+}
+
+// MARK: - Scroll Detection for macOS 14
+
+private struct ScrollDetectionView: View {
+    let onScrollingChanged: (Bool) -> Void
+    @State private var lastOffset: CGFloat = 0
+    @State private var scrollTimer: Timer?
+    
+    var body: some View {
+        GeometryReader { geometry in
+            Color.clear
+                .preference(
+                    key: ScrollOffsetKey.self,
+                    value: geometry.frame(in: .named("scroll")).origin.y
+                )
+        }
+        .onPreferenceChange(ScrollOffsetKey.self) { newOffset in
+            if abs(newOffset - lastOffset) > 1 {
+                onScrollingChanged(true)
+                lastOffset = newOffset
+                
+                scrollTimer?.invalidate()
+                scrollTimer = Timer.scheduledTimer(withTimeInterval: 0.15, repeats: false) { _ in
+                    onScrollingChanged(false)
+                }
+            }
+        }
+    }
+}
+
+private struct ScrollOffsetKey: PreferenceKey {
+    static var defaultValue: CGFloat = 0
+    static func reduce(value: inout CGFloat, nextValue: () -> CGFloat) {
+        value = nextValue()
+    }
+}
+
+// MARK: - Image Cache
+
+private class RenderedImageCache {
+    static let shared = RenderedImageCache()
+    private let cache = NSCache<NSString, NSImage>()
+    
+    init() {
+        cache.countLimit = 1000
+    }
+    
+    func getImage(for entity: any Entity) -> NSImage? {
+        let key = "\(entity.id.uuidString)-rendered" as NSString
+        
+        if let cached = cache.object(forKey: key) {
+            return cached
+        }
+        
+        guard let artworkData = entity.artworkLarge else { return nil }
+        
+        let renderedImage = createRenderedImage(from: artworkData)
+        
+        if let image = renderedImage {
+            cache.setObject(image, forKey: key)
+        }
+        
+        return renderedImage
+    }
+    
+    private func createRenderedImage(from data: Data) -> NSImage? {
+        guard let originalImage = NSImage(data: data) else { return nil }
+        
+        let size = NSSize(width: ViewDefaults.gridArtworkSize, height: ViewDefaults.gridArtworkSize)
+        let renderedImage = NSImage(size: size)
+        
+        renderedImage.lockFocus()
+        
+        let rect = NSRect(origin: .zero, size: size)
+        let path = NSBezierPath(roundedRect: rect, xRadius: 8, yRadius: 8)
+        path.addClip()
+        
+        originalImage.draw(in: rect, from: .zero, operation: .copy, fraction: 1.0)
+        
+        renderedImage.unlockFocus()
+        
+        return renderedImage
+    }
+}
+
+// MARK: - Simplified Grid Item
 private struct EntityGridItem<T: Entity>: View {
     let entity: T
     let isHovered: Bool
@@ -95,107 +192,75 @@ private struct EntityGridItem<T: Entity>: View {
     let onSelect: () -> Void
     let onHover: (Bool) -> Void
 
-    @State private var artworkImage: NSImage?
+    @State private var renderedImage: NSImage?
 
     var body: some View {
         VStack(spacing: 8) {
-            artworkSection
-            textSection
+            Group {
+                if let image = renderedImage {
+                    Image(nsImage: image)
+                        .frame(width: ViewDefaults.gridArtworkSize, height: ViewDefaults.gridArtworkSize)
+                        .shadow(color: .black.opacity(0.2), radius: 10, x: 0, y: 5)
+                } else {
+                    Rectangle()
+                        .fill(Color.gray.opacity(0.2))
+                        .frame(width: ViewDefaults.gridArtworkSize, height: ViewDefaults.gridArtworkSize)
+                        .overlay(
+                            Image(systemName: Icons.entityIcon(for: entity))
+                                .font(.system(size: 48))
+                                .foregroundColor(.gray)
+                        )
+                        .cornerRadius(8)
+                        .shadow(color: .black.opacity(0.2), radius: 10, x: 0, y: 5)
+                }
+            }
+            .onAppear {
+                loadArtwork()
+            }
+            
+            VStack(alignment: .leading, spacing: 2) {
+                Text(entity.name)
+                    .font(.system(size: 13, weight: .semibold))
+                    .lineLimit(1)
+                    .foregroundColor(.primary)
+
+                if let subtitle = entity.subtitle {
+                    Text(subtitle)
+                        .font(.system(size: 11))
+                        .foregroundColor(.secondary)
+                        .lineLimit(1)
+                }
+
+                if entity is AlbumEntity {
+                    Text("\(entity.trackCount) \(entity.trackCount == 1 ? "song" : "songs")")
+                        .font(.system(size: 11))
+                        .foregroundColor(.secondary)
+                        .lineLimit(1)
+                }
+            }
+            .frame(width: ViewDefaults.gridArtworkSize, alignment: .leading)
         }
         .padding(8)
-        .background(backgroundView)
+        .background(
+            RoundedRectangle(cornerRadius: 10)
+                .fill(isHovered ? Color(NSColor.selectedContentBackgroundColor).opacity(0.15) : Color.clear)
+                .animation(
+                    isScrolling ? .none : .easeInOut(duration: 0.08),
+                    value: isHovered
+                )
+        )
         .contentShape(Rectangle())
         .onTapGesture(perform: onSelect)
         .onHover(perform: onHover)
-        .animation(.easeInOut(duration: AnimationDuration.quickDuration), value: isHovered)
     }
     
-    @ViewBuilder
-    private var artworkSection: some View {
-        ZStack {
-            RoundedRectangle(cornerRadius: 8)
-                .fill(Color.gray.opacity(0.2))
-                .aspectRatio(1, contentMode: .fit)
-
-            if let image = artworkImage {
-                Image(nsImage: image)
-                    .resizable()
-                    .aspectRatio(contentMode: .fill)
-                    .frame(width: ViewDefaults.gridArtworkSize, height: ViewDefaults.gridArtworkSize)
-                    .clipped()
-                    .cornerRadius(8)
-            } else {
-                VStack(spacing: 8) {
-                    Image(systemName: Icons.entityIcon(for: entity))
-                        .font(.system(size: 48))
-                        .foregroundColor(.gray)
-
-                    Text(entity.name)
-                        .font(.system(size: 11, weight: .medium))
-                        .foregroundColor(.gray)
-                        .lineLimit(2)
-                        .multilineTextAlignment(.center)
-                        .padding(.horizontal, 8)
-                }
+    private func loadArtwork() {
+        DispatchQueue.global(qos: .userInitiated).async {
+            let image = RenderedImageCache.shared.getImage(for: entity)
+            
+            DispatchQueue.main.async {
+                self.renderedImage = image
             }
         }
-        .frame(width: ViewDefaults.gridArtworkSize, height: ViewDefaults.gridArtworkSize)
-        .task(id: entity.id) {
-            await loadEntityArtwork()
-        }
-        .onDisappear {
-            artworkImage = nil
-        }
-    }
-    
-    private var textSection: some View {
-        VStack(alignment: .leading, spacing: 2) {
-            Text(entity.name)
-                .font(.system(size: 13, weight: .semibold))
-                .lineLimit(1)
-                .foregroundColor(.primary)
-                .help(entity.name)
-
-            if let subtitle = entity.subtitle {
-                Text(subtitle)
-                    .font(.system(size: 11))
-                    .foregroundColor(.secondary)
-                    .lineLimit(1)
-                    .help(subtitle)
-            }
-
-            if entity is AlbumEntity {
-                Text("\(entity.trackCount) \(entity.trackCount == 1 ? "song" : "songs")")
-                    .font(.system(size: 11))
-                    .foregroundColor(.secondary)
-                    .lineLimit(1)
-            }
-        }
-        .frame(width: ViewDefaults.gridArtworkSize, alignment: .leading)
-    }
-
-    private var backgroundView: some View {
-        RoundedRectangle(cornerRadius: 10)
-            .fill(isHovered ? Color(NSColor.selectedContentBackgroundColor).opacity(0.15) : Color.clear)
-            .animation(.easeInOut(duration: AnimationDuration.quickDuration), value: isHovered)
-    }
-
-    private func loadEntityArtwork() async {
-        let delay = isScrolling ? TimeConstants.oneFiftyMilliseconds : TimeConstants.fiftyMilliseconds
-        
-        await loadEntityArtworkAsync(
-            from: entity.artworkLarge,
-            into: $artworkImage,
-            delay: delay
-        )
-    }
-}
-
-// MARK: - Supporting Types
-
-private struct ScrollOffsetPreferenceKey: PreferenceKey {
-    static var defaultValue: CGFloat = 0
-    static func reduce(value: inout CGFloat, nextValue: () -> CGFloat) {
-        value = nextValue()
     }
 }
